Conflicts:
	gatling-core/src/test/scala/com/excilys/ebi/gatling/core/result/writer/ConsoleDataWriterSpec.scala

+++ gatling-charts/src/main/scala/com/excilys/ebi/gatling/charts/component/ComponentLibrary.scala
@@ -44,21 +44,21 @@ object ComponentLibrary extends Logging {
 
 abstract class ComponentLibrary {
 
-	def getAllSessionsJs(series: Series[Long, Int]): String
+	def getAllSessionsJs(series: Series[Long, Long]): String
 
-	def getActiveSessionsChartComponent(series: Seq[Series[Long, Int]]): Component
+	def getActiveSessionsChartComponent(series: Seq[Series[Long, Long]]): Component
 
-	def getRequestsChartComponent(allRequests: Series[Long, Int], failedRequests: Series[Long, Int], succeededRequests: Series[Long, Int], pieSeries: Series[String, Int]): Component
+	def getRequestsChartComponent(allRequests: Series[Long, Long], failedRequests: Series[Long, Long], succeededRequests: Series[Long, Long], pieSeries: Series[String, Long]): Component
 
-	def getTransactionsChartComponent(allTransactions: Series[Long, Int], failedTransactions: Series[Long, Int], succeededTransactions: Series[Long, Int], pieSeries: Series[String, Int]): Component
+	def getTransactionsChartComponent(allTransactions: Series[Long, Long], failedTransactions: Series[Long, Long], succeededTransactions: Series[Long, Long], pieSeries: Series[String, Long]): Component
 
 	def getRequestDetailsResponseTimeChartComponent(responseTimesSuccess: Series[Long, Long], responseTimesFailures: Series[Long, Long]): Component
 
-	def getRequestDetailsResponseTimeDistributionChartComponent(responseTimesSuccess: Series[Long, Int], responseTimesFailures: Series[Long, Int]): Component
+	def getRequestDetailsResponseTimeDistributionChartComponent(responseTimesSuccess: Series[Long, Long], responseTimesFailures: Series[Long, Long]): Component
 
 	def getRequestDetailsLatencyChartComponent(latencySuccess: Series[Long, Long], latencyFailures: Series[Long, Long]): Component
 
-	def getRequestDetailsScatterChartComponent(successData: Series[Int, Long], failuresData: Series[Int, Long]): Component
+	def getRequestDetailsScatterChartComponent(successData: Series[Long, Long], failuresData: Series[Long, Long]): Component
 
 	def getRequestDetailsIndicatorChartComponent: Component
 }
